+++ pkg/util/response.go
@@ -2,6 +2,8 @@ package util
 
 import "github.com/wklken/apisix-go/pkg/json"
 
+// TODO: use a pool here?
+
 func BuildMessageResponse(message string) string {
 	body, _ := json.Marshal(map[string]string{
 		"message": message,